chore: 콘솔 관련 파일 이동
